+++ CHANGELOG.md
@@ -9,8 +9,10 @@ An example of configuration:
 ```toml
 [navigation]
 mode = "Breadcrumb"
+use-current-path = true
 ```
 
+- Introduce `navigation.use-current-path` which sets if a tab/breacrumb should be open from the current context path.
 - Fix rendering unicode with 1 width glyphs (fix [#160]https://github.com/raphamorim/rio/issues/160)).
 - Increased max tabs from 9 to 72.
 - Default colors `tab` and `tab-active` has changed.